+++ src/wasm-vector.h
@@ -108,7 +108,7 @@
 
 #define WASM_DESTROY_VECTOR_AND_ELEMENTS(allocator, v, name) \
   {                                                          \
-    int i;                                                   \
+    size_t i;                                                \
     for (i = 0; i < (v).size; ++i)                           \
       wasm_destroy_##name(allocator, &((v).data[i]));        \
     wasm_destroy_##name##_vector(allocator, &(v));           \
